requirements.txt
anthropic
audio_recorder_streamlit
gTTS
langchain
langchain_community
langchain_core
langchain_google_genai
langchain_groq
openai
Pillow
protobuf
PyPDF2
Pyrebase4
pyreadline3
python-dotenv
reportlab
Requests
SpeechRecognition
st_annotated_text
streamlit
streamlit_lottie
streamlit_option_menu
tqdm
streamlit
setuptools
faiss-cpu
firebase-admin
pyrebase4 
oauth2client
tensorflow
pdfplumber
